Move KType extensions to new KTypes.kt file
